Separated out serializer fields that aren't needed immediately
