+++ contentcuration/contentcuration/static/js/edit_channel/export/hbtemplates/export_modal.handlebars
@@ -27,7 +27,7 @@
                         </button>
                         <span id="export_totals" class="pull-right">
                             {{#format_count "Resource" node.metadata.resource_count}}{{/format_count}}
-                            <em class="annotation" id="export_size">(Calculating size...)</em>
+                            <em class="annotation" id="export_size">(Calculating...)</em>
                         </span>
                         <a class="action-text" data-dismiss="modal">
                             <span>CANCEL</span>
