+++ contentcuration/contentcuration/static/js/edit_channel/import/hbtemplates/import_dialog.handlebars
@@ -7,5 +7,8 @@
 	<button disabled class="action-button pull-right" id="import_content_submit">
 		Select content to import...
 	</button>
-	<span id="import_file_count" class="pull-right"><em>No files selected</em></span>
+	<span id="import_file_metadata" class="pull-right">
+		<span id="import_file_count">0 Topics, 0 Resources</span>
+		(<em id="import_file_size">0B</em>)
+	</span>
 </div>
