+++ contentcuration/contentcuration/static/js/edit_channel/tree_edit/hbtemplates/content_list_item.handlebars
@@ -21,13 +21,15 @@
 				{{#if time}}&nbsp;<span class="glyphicon glyphicon-time" aria-hidden="true"></span>{{time}}{{/if}}
 				{{#if isfolder}}&nbsp;{{#format_count "Resource" node.metadata.resource_count}}{{/format_count}}{{/if}}
 				&nbsp;
-				{{#if isexercise}}
-					<span class="glyphicon glyphicon-check" aria-hidden="true"></span>
-					{{#format_count "Question" num_questions}}{{/format_count}}
-				{{else}}
-					<span class="glyphicon glyphicon-file" aria-hidden="true"></span>
-					{{#format_file_size node.metadata.resource_size}}{{/format_file_size}}
-				{{/if}}
+				{{#unless isfolder}}
+					{{#if isexercise}}
+						<span class="glyphicon glyphicon-check" aria-hidden="true"></span>
+						{{#format_count "Question" num_questions}}{{/format_count}}
+					{{else}}
+						<span class="glyphicon glyphicon-file" aria-hidden="true"></span>
+						{{#format_file_size node.metadata.resource_size}}{{/format_file_size}}
+					{{/if}}
+				{{/unless}}
 				&nbsp;
 				{{#if node.changed}}
 					<em class="unpublished">{{#if node.published}}Updated{{else}}New{{/if}}</em>
