+++ contentcuration/contentcuration/static/less/import.less
@@ -129,8 +129,8 @@
 		#import_content_submit{
 			margin-right:20px;
 		}
-		#import_file_count{
-			margin-right:20px;
+		#import_file_metadata{
+			padding-right:20px;
 			font-size:12pt;
 			margin-top:2px;
 		}
