+++ contentcuration/contentcuration/views.py
@@ -18,7 +18,7 @@ from django.core.management import call_command
 from django.core.exceptions import ObjectDoesNotExist
 from django.core.context_processors import csrf
 from django.db import transaction
-from django.db.models import Q, Case, When, Value, IntegerField, Max
+from django.db.models import Q, Case, When, Value, IntegerField, Max, Sum
 from django.core.urlresolvers import reverse_lazy
 from django.core.files import File as DjFile
 from rest_framework.renderers import JSONRenderer
@@ -33,12 +33,6 @@ from rest_framework.decorators import api_view, authentication_classes, permissi
 from pressurecooker.videos import guess_video_preset_by_resolution, extract_thumbnail_from_video, compress_video
 from pressurecooker.images import create_tiled_image
 from pressurecooker.encodings import write_base64_to_file
-
-def get_nodes_by_ids(request):
-    if request.method == 'POST':
-        nodes = ContentNode.objects.prefetch_related('files').prefetch_related('assessment_items')\
-                .prefetch_related('tags').prefetch_related('children').filter(pk__in=json.loads(request.body))
-        return HttpResponse(JSONRenderer().render(ContentNodeSerializer(nodes, many=True).data))
 
 def base(request):
     if not check_supported_browsers(request.META.get('HTTP_USER_AGENT')):
@@ -142,238 +136,6 @@ def channel_view_only(request, channel_id):
         return redirect(reverse_lazy('unauthorized'))
 
     return channel_page(request, channel)
-
-def file_upload(request):
-    if request.method == 'POST':
-        preset = FormatPreset.objects.get(id=request.META.get('HTTP_PRESET'))
-        #Implement logic for switching out files without saving it yet
-        filename, ext = os.path.splitext(request.FILES.values()[0]._name)
-        size = request.FILES.values()[0]._size
-        file_object = File(file_size=size, file_on_disk=DjFile(request.FILES.values()[0]), file_format_id=ext[1:], original_filename=request.FILES.values()[0]._name, preset=preset)
-        file_object.save()
-        return HttpResponse(json.dumps({
-            "success": True,
-            "filename": str(file_object),
-            "file": JSONRenderer().render(FileSerializer(file_object).data)
-        }))
-
-def file_create(request):
-    if request.method == 'POST':
-        original_filename, ext = os.path.splitext(request.FILES.values()[0]._name)
-        size = request.FILES.values()[0]._size
-        presets = FormatPreset.objects.filter(allowed_formats__extension__contains=ext[1:])
-        kind = presets.first().kind
-        preferences = json.loads(request.user.preferences)
-        author = preferences.get('author') if isinstance(preferences.get('author'), basestring) else request.user.get_full_name()
-        license = License.objects.filter(license_name=preferences.get('license')).first() # Use filter/first in case preference hasn't been set
-        license_id = license.pk if license else settings.DEFAULT_LICENSE
-        new_node = ContentNode(title=original_filename, kind=kind, license_id=license_id, author=author, copyright_holder=preferences.get('copyright_holder'))
-        if license.license_name == licenses.SPECIAL_PERMISSIONS:
-            new_node.license_description = preferences.get('license_description')
-        new_node.save()
-        file_object = File(file_on_disk=DjFile(request.FILES.values()[0]), file_format_id=ext[1:], original_filename=request.FILES.values()[0]._name, contentnode=new_node, file_size=size)
-        file_object.save()
-        if kind.pk == content_kinds.VIDEO:
-            file_object.preset_id = guess_video_preset_by_resolution(str(file_object.file_on_disk))
-        elif presets.filter(supplementary=False).count() == 1:
-            file_object.preset = presets.filter(supplementary=False).first()
-
-        file_object.save()
-
-        try:
-            if preferences.get('auto_derive_video_thumbnail') and new_node.kind_id == content_kinds.VIDEO \
-                or preferences.get('auto_derive_audio_thumbnail') and new_node.kind_id == content_kinds.AUDIO \
-                or preferences.get('auto_derive_html5_thumbnail') and new_node.kind_id == content_kinds.HTML5 \
-                or preferences.get('auto_derive_document_thumbnail') and new_node.kind_id == content_kinds.DOCUMENT:
-                generate_thumbnail_from_node(new_node, set_node=True)
-        except Exception:
-            pass
-
-        return HttpResponse(json.dumps({
-            "success": True,
-            "node": JSONRenderer().render(ContentNodeSerializer(new_node).data)
-        }))
-
-def generate_thumbnail(request):
-    logging.debug("Entering the generate_thumbnail endpoint")
-
-    if request.method != 'POST':
-        raise HttpResponseBadRequest("Only POST requests are allowed on this endpoint.")
-    else:
-        data = json.loads(request.body)
-        node = ContentNode.objects.get(pk=data["node_id"])
-
-        thumbnail_object = generate_thumbnail_from_node(node)
-
-        return HttpResponse(json.dumps({
-            "success": True,
-            "file": JSONRenderer().render(FileSerializer(thumbnail_object).data),
-            "path": generate_storage_url(str(thumbnail_object)),
-        }))
-
-def thumbnail_upload(request):
-    if request.method == 'POST':
-        fobj = request.FILES.values()[0]
-        formatted_filename = write_file_to_storage(fobj)
-
-        return HttpResponse(json.dumps({
-            "success": True,
-            "formatted_filename": formatted_filename,
-            "file":  None,
-            "path": generate_storage_url(formatted_filename),
-        }))
-
-def image_upload(request):
-    if request.method == 'POST':
-        name, ext = os.path.splitext(request.FILES.values()[0]._name) # gets file extension without leading period
-        file_object = File(contentnode_id=request.META.get('HTTP_NODE'),original_filename=name, preset_id=request.META.get('HTTP_PRESET'), file_on_disk=DjFile(request.FILES.values()[0]), file_format_id=ext[1:])
-        file_object.save()
-        return HttpResponse(json.dumps({
-            "success": True,
-            "file": JSONRenderer().render(FileSerializer(file_object).data),
-            "path": generate_storage_url(str(file_object)),
-        }))
-
-def exercise_image_upload(request):
-    if request.method == 'POST':
-        ext = os.path.splitext(request.FILES.values()[0]._name)[1][1:] # gets file extension without leading period
-        file_object = File(preset_id=format_presets.EXERCISE_IMAGE, file_on_disk=DjFile(request.FILES.values()[0]), file_format_id=ext)
-        file_object.save()
-        return HttpResponse(json.dumps({
-            "success": True,
-            "formatted_filename": exercises.CONTENT_STORAGE_FORMAT.format(str(file_object)),
-            "file_id": file_object.pk,
-            "path": generate_storage_url(str(file_object)),
-        }))
-
-def duplicate_nodes(request):
-    logging.debug("Entering the copy_node endpoint")
-
-    if request.method != 'POST':
-        raise HttpResponseBadRequest("Only POST requests are allowed on this endpoint.")
-    else:
-        data = json.loads(request.body)
-
-        try:
-            nodes = data["nodes"]
-            sort_order = data.get("sort_order") or 1
-            target_parent = data["target_parent"]
-            channel_id = data["channel_id"]
-            new_nodes = []
-
-            with transaction.atomic():
-                for node_data in nodes:
-                    new_node = _duplicate_node(node_data['id'], sort_order=sort_order, parent=target_parent, channel_id=channel_id)
-                    new_nodes.append(new_node.pk)
-                    sort_order+=1
-
-        except KeyError:
-            raise ObjectDoesNotExist("Missing attribute from data: {}".format(data))
-
-        return HttpResponse(json.dumps({
-            "success": True,
-            "node_ids": " ".join(new_nodes)
-        }))
-
-def _duplicate_node(node, sort_order=None, parent=None, channel_id=None):
-    if isinstance(node, int) or isinstance(node, basestring):
-        node = ContentNode.objects.get(pk=node)
-
-    original_channel = node.get_original_node().get_channel() if node.get_original_node() else None
-
-    new_node = ContentNode.objects.create(
-        title=node.title,
-        description=node.description,
-        kind=node.kind,
-        license=node.license,
-        parent=ContentNode.objects.get(pk=parent) if parent else None,
-        sort_order=sort_order or node.sort_order,
-        copyright_holder=node.copyright_holder,
-        changed=True,
-        original_node=node.original_node or node,
-        cloned_source=node,
-        original_channel_id = node.original_channel_id or original_channel.id if original_channel else None,
-        source_channel_id = node.get_channel().id if node.get_channel() else None,
-        original_source_node_id = node.original_source_node_id or node.node_id,
-        source_node_id = node.node_id,
-        author=node.author,
-        content_id=node.content_id,
-        extra_fields=node.extra_fields,
-    )
-
-    # add tags now
-    for tag in node.tags.all():
-        new_tag, is_new = ContentTag.objects.get_or_create(
-            tag_name=tag.tag_name,
-            channel_id=channel_id,
-        )
-        new_node.tags.add(new_tag)
-
-    # copy file object too
-    for fobj in node.files.all():
-        duplicate_file(fobj, node=new_node)
-
-    # copy assessment item object too
-    for aiobj in node.assessment_items.all():
-        aiobj_copy = copy.copy(aiobj)
-        aiobj_copy.id = None
-        aiobj_copy.contentnode = new_node
-        aiobj_copy.save()
-        for fobj in aiobj.files.all():
-            duplicate_file(fobj, assessment_item=aiobj_copy)
-
-    for c in node.children.all():
-        _duplicate_node(c, parent=new_node.id)
-
-    return new_node
-
-def move_nodes(request):
-    logging.debug("Entering the move_nodes endpoint")
-
-    if request.method != 'POST':
-        raise HttpResponseBadRequest("Only POST requests are allowed on this endpoint.")
-    else:
-        data = json.loads(request.body)
-
-        try:
-            nodes = data["nodes"]
-            target_parent = ContentNode.objects.get(pk=data["target_parent"])
-            channel_id = data["channel_id"]
-            min_order = data.get("min_order") or 0
-            max_order = data.get("max_order") or min_order + len(nodes)
-
-        except KeyError:
-            raise ObjectDoesNotExist("Missing attribute from data: {}".format(data))
-
-        all_ids = []
-        with transaction.atomic():
-            for n in nodes:
-                min_order = min_order + float(max_order - min_order) / 2
-                node = ContentNode.objects.get(pk=n['id'])
-                _move_node(node, parent=target_parent, sort_order=min_order, channel_id=channel_id)
-                all_ids.append(n['id'])
-
-        serialized = ContentNodeSerializer(ContentNode.objects.filter(pk__in=all_ids), many=True).data
-        return HttpResponse(JSONRenderer().render(serialized))
-
-def _move_node(node, parent=None, sort_order=None, channel_id=None):
-    node.parent = parent
-    node.sort_order = sort_order
-    node.changed = True
-    descendants = node.get_descendants(include_self=True)
-    node.save()
-
-    for tag in ContentTag.objects.filter(tagged_content__in=descendants).distinct():
-        # If moving from another channel
-        if tag.channel_id != channel_id:
-            t, is_new = ContentTag.objects.get_or_create(tag_name=tag.tag_name, channel_id=channel_id)
-
-            # Set descendants with this tag to correct tag
-            for n in descendants.filter(tags=tag):
-                n.tags.remove(tag)
-                n.tags.add(t)
-
-    return node
 
 @csrf_exempt
 def publish_channel(request):
